+++ game/gamestates/gameover.lua
@@ -57,6 +57,8 @@ function state:update(dt)
         Gamestate.switch(GS.MENU)
     end
 
+    Util.updateTimers(dt)
+
 end
 
 function state:draw()