+++ docs/CHANGELOG.md
@@ -2,10 +2,13 @@
 THE HISTORY OF PSYCHO: THE BALL
 ===============================
 
+#### (xx/xx/2016) Version 0.0.0.5    -  Snowballing
 
-#### (xx/xx/2016) Version 0.0.0.1    -  From Ashes
-
+#### (29/08/2016) Version 0.0.0.1    -  From Ashes
 
 - Using [*STEAMING*](https://github.com/uspgamedev/STEAMING) template ver. 1.0.0, started making PsyChO.
 - Main documents ready (README, DEVLOG, LICENSE and CHANGELOG)
 - Started using issues on [*github repository*](https://github.com/uspgamedev/Project-Telos)
+- Psycho character can now move, shoot, and kill enemies
+- Has a simple menu, game, pause and gameover screen
+- Button interactions working (used for changing gamestates)